services/reporting/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka & Database
kafka-python==2.0.2
psycopg2-binary==2.9.9

# OpenAI
openai==1.10.0

# Data Processing
pandas==2.1.4
jinja2==3.1.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Monitoring
prometheus-client==0.19.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
